More optimization and multithreading
